RAG and Agentic AI/trading_agents/requirements.txt
crewai==0.140.0
crewai-tools==0.49.0
python-dotenv==1.1.1
yfinance==0.2.64